+++ example/gen_example.dart
@@ -4,4 +4,5 @@ void main() {
   var gen = Generate();
   print('bool: ${gen.boolean()}');
   print('title: ${gen.title()}');
+  print('int: ${gen.integer()}');
 }