+++ lib/src/gen_base.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 // TODO: Wrong place
 var text =
-    'Teoria da Relatividade é a denominação dada ao conjunto de duas teorias científicas: a Relatividade Restrita e a Relatividade Geral. A Relatividade Especial é uma teoria publicada no ano de 1905 por Albert Einstein, concluindo estudos precedentes do físico neerlandês Hendrik Lorentz, entre outros. Ela substitui os conceitos independentes de espaço e tempo da Teoria de Newton pela ideia de espaço-tempo como uma entidade geométrica unificada. O espaço-tempo na relatividade especial consiste de uma variedade diferenciável de 4 dimensões, três espaciais e uma temporal, munida de uma métrica pseudo-riemanniana, o que permite que noções de geometria possam ser utilizadas. É nessa teoria, também, que surge a ideia de velocidade da luz invariante. O termo especial é usado porque ela é um caso particular do princípio da relatividade em que efeitos da gravidade são ignorados. Dez anos após a publicação da teoria especial, Einstein publicou a Teoria Geral da Relatividade, que é a versão mais ampla da teoria, em que os efeitos da gravitação são integrados, surgindo a noção de espaço-tempo curvo. Cor é como o cérebro interpreta os sinais eletro nervosos vindos do olho, resultantes da reemissão da luz vinda de um objeto que foi emitida por uma fonte luminosa por meio de ondas eletromagnéticas; e que corresponde à parte do espectro eletromagnético que é visível. A Cor não é um fenômeno físico. Um mesmo comprimento de onda pode ser percebido diferentemente por diferentes pessoas, ou seja, cor é um fenômeno fisiológico, de caráter subjetivo e individual.';
+    "Teoria da Relatividade é a denominação dada ao conjunto de duas teorias científicas: a Relatividade Restrita e a Relatividade Geral. A Relatividade Especial é uma teoria publicada no ano de 1905 por Albert Einstein, concluindo estudos precedentes do físico neerlandês Hendrik Lorentz, entre outros. Ela substitui os conceitos independentes de espaço e tempo da Teoria de Newton pela ideia de espaço-tempo como uma entidade geométrica unificada. O espaço-tempo na relatividade especial consiste de uma variedade diferenciável de 4 dimensões, três espaciais e uma temporal, munida de uma métrica pseudo-riemanniana, o que permite que noções de geometria possam ser utilizadas. É nessa teoria, também, que surge a ideia de velocidade da luz invariante. O termo especial é usado porque ela é um caso particular do princípio da relatividade em que efeitos da gravidade são ignorados. Dez anos após a publicação da teoria especial, Einstein publicou a Teoria Geral da Relatividade, que é a versão mais ampla da teoria, em que os efeitos da gravitação são integrados, surgindo a noção de espaço-tempo curvo. Cor é como o cérebro interpreta os sinais eletro nervosos vindos do olho, resultantes da reemissão da luz vinda de um objeto que foi emitida por uma fonte luminosa por meio de ondas eletromagnéticas; e que corresponde à parte do espectro eletromagnético que é visível. A Cor não é um fenômeno físico. Um mesmo comprimento de onda pode ser percebido diferentemente por diferentes pessoas, ou seja, cor é um fenômeno fisiológico, de caráter subjetivo e individual. Rey Skywalker, nascida Rey Palpatine, era uma Humana catadora de sucata nativa de Jakku que descobriu ser sensível à Força durante sua busca ao lendário Mestre Jedi Luke Skywalker a fim de trazer uma nova esperança à galáxia à beira da guerra. Rey cresceu sozinha no mundo desértico de Jakku, esperando sua família retornar, até encontrar BB-8, droide que carregava informações sobre o paradeiro de Skywalker. Juntamente com Finn, um desertor da Primeira Ordem, Han Solo e Chewbacca, ela embarcou em uma jornada para entregar o mapa a Skywalker para General Leia Organa e para a Resistência. No caminho, ela descobriu o sabre de luz de Skywalker e sua sensibilidade à Força despertou. Logo depois, ela acabou sendo capturada por Kylo Ren, nascido Ben Solo, filho de Leia e Han, contra quem Rey duelou na Base Starkiller logo antes da Resistência destruir a superarma. Depois de chegar à base da Resistência em D'Qar, Leia deu a Rey a missão de encontrar Luke com base no mapa e apresentar a ele o seu antigo sabre de luz como um símbolo da única esperança da galáxia: a volta dos Jedi.";
 
 class Generate {
   final _r = Random();
@@ -11,25 +11,36 @@ class Generate {
       ? throw RangeError('argument \'likelihood\' accepts values from 0 to 100')
       : _r.nextDouble() * 100 < likelihood;
 
-  int integer([max = 999]) => _r.nextInt(max);
+  int integer([max = 999, min = 0]) {
+    var i = _r.nextInt(max);
+    if (i < min) i = min;
+    return i;
+  }
+
+  List<String> words() => text.split(' ')..shuffle(_r);
+
+  String firstWord() => words()
+      .firstWhere((s) => RegExp('([A-Z][a-z0-9]+)+').hasMatch(s))
+      .replaceAll('.', '');
+
+  String lastWord() =>
+      words().firstWhere((s) => RegExp('([+.])+').hasMatch(s)).toLowerCase();
 
   String title() {
-    var words = integer(8);
-    var data = text.split(' ')..shuffle(_r);
-    var first = data.firstWhere((d) => RegExp('([A-Z][a-z0-9]+)+').hasMatch(d));
-    var last = data.firstWhere((d) => RegExp('([.]+)+').hasMatch(d));
+    var data = words();
+    var first = firstWord();
     var start = integer(data.length);
-    var end = start + words;
+    var end = start + integer(7, 1);
     if (end > data.length) end = data.length;
+
     return [
-      first.trim(),
-      data
+      first,
+      ...data
           .getRange(start, end)
-          .where((s) => RegExp('^[^.]+\$').hasMatch(s))
+          .map((s) => s.length > 1 ? s : s.toLowerCase())
           .toSet()
-          .join(' ')
-          .toLowerCase(),
-      last.trim().toLowerCase(),
+          .where((s) => !s.contains(first) && RegExp('^[^.]+\$').hasMatch(s)),
+      lastWord().replaceAll('.', ''),
     ].join(' ').trim();
   }
 }