+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.1+3
+
+- Improved `title()` return
+- Added `words()` generator`
+
 ## 0.0.1+2
 
 - Fix example